requirements.txt
streamlit
openai
pandas
python-dotenv
PyPDF2
